Update InternetArchiveBot
1) Some further memory refinements
2) More advancements in controlled debugging
3) Added memory, so bot may resume where it leaves off
4) Refinements to the function that determines when the link was add.
It's still slow, but it doesn't blow up the memory for large articles,
thanks to the suggestion of implementing a binary search.  If tied to a
wiki database with text and revision tables, the search speed will go
up.
5) Added hashtags to edit summaries as requested.
6) Fixed a minor bug in getTimeAdded function resulting in an occasional
incorrect timestamp.
7) This is last version that will not be dependent on a local database.
The next script release will require a DB or the bot will die.