Update IABot
Fix a minor bug causing ridiculous amounts of errors to be spewed, with
no functional consequence to operations.
Fixed a critical bug that caused the bot to ignore the error flag that
is raised when confronted with confusing/conflicting data in a ref,
resulting reference blanking.
Updated archive validator to ignore URLs to the archive sites, when they
don't point to a specific snapshot, but force the presence of a snapshot
when found in the archive templates/parameters.