Update IABot to v1.2.6
Fix a bug where HTML comments were misparsed, in an attempt to preserve
them in templates.  Now they get purged when parsing the template.
Fix a bug where the absence of deadurl parameters in cite templates
lead to the bot misinterpreting the source as alive when archive URLs
were present.